test(unit): migrate the base questions tests from mocha to vitest
for #767
