+++ .eslintrc.yml
@@ -1,6 +1,7 @@
 extends:
   - '@form8ion'
   - '@form8ion/mocha'
+  - '@form8ion/vitest'
   - '@form8ion/cucumber'
 
 overrides:
