+++ package.json
@@ -78,6 +78,7 @@
     "@form8ion/eslint-config": "8.0.0-beta.1",
     "@form8ion/eslint-config-cucumber": "1.4.1",
     "@form8ion/eslint-config-mocha": "3.0.6",
+    "@form8ion/eslint-config-vitest": "1.1.0",
     "@form8ion/remark-lint-preset": "6.0.7",
     "@rollup/plugin-babel": "7.0.0",
     "@travi/any": "3.3.0",
